Add stremio login
This is an initial POC for login with a Stremio account. When the user clicks log in, a single API request is sent which logs the user in. If successful, the authentication key is read and 'copied over' to the authentication key field so the addon can be used as usual. If there's an error, it's logged to console and an alert is shown to the user.